requirements.txt
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
protobuf==3.12.2
PyMySQL==0.9.3
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==0.15.2
